Create License
All rights reserved © 2025 Javier Linares.

This project is currently in active development. No license is granted for reproduction, distribution, or modification of any part of this codebase.

You may view the source code for personal inspection only.  
Commercial use, forking, copying, or reuse of this code is not permitted without written permission from the author.